requirements.txt
django==1.6.7   # Django main package
djangorestframework==2.3.14     # API
south==1.0      # For database migrations